Update QLM JTAG code to handle fields that are wider than 32 bits. Fields up to 64bits should now be supported.
git-svn-id: http://casvn/svn/octeon/bdk/trunk@60677 c83883e8-9f28-0410-baf1-f2025811e16a

+++ libbdk-hal/bdk-qlm-jtag.h
@@ -3,7 +3,7 @@
     the higher level functions in bdk-qlm */
 
 extern void __bdk_qlm_jtag_init(void);
-extern uint32_t __bdk_qlm_jtag_shift(int qlm, int bits, uint32_t data);
+extern uint64_t __bdk_qlm_jtag_shift(int qlm, int bits, uint64_t data);
 extern void __bdk_qlm_jtag_shift_zeros(int qlm, int bits);
 extern void __bdk_qlm_jtag_update(int qlm);
 extern void __bdk_qlm_jtag_capture(int qlm);
